[FLINK-11494][runtime] Remove WebRuntimeMonitor
