+++ flink-runtime/src/main/java/org/apache/flink/runtime/webmonitor/WebMonitorUtils.java
@@ -22,22 +22,15 @@ import org.apache.flink.api.common.time.Time;
 import org.apache.flink.configuration.Configuration;
 import org.apache.flink.configuration.WebOptions;
 import org.apache.flink.core.fs.Path;
-import org.apache.flink.runtime.concurrent.ScheduledExecutor;
 import org.apache.flink.runtime.dispatcher.DispatcherGateway;
 import org.apache.flink.runtime.execution.ExecutionState;
 import org.apache.flink.runtime.executiongraph.AccessExecutionGraph;
 import org.apache.flink.runtime.executiongraph.AccessExecutionJobVertex;
 import org.apache.flink.runtime.executiongraph.AccessExecutionVertex;
-import org.apache.flink.runtime.highavailability.HighAvailabilityServices;
 import org.apache.flink.runtime.jobgraph.JobStatus;
-import org.apache.flink.runtime.jobmaster.JobManagerGateway;
-import org.apache.flink.runtime.leaderretrieval.LeaderRetrievalService;
 import org.apache.flink.runtime.messages.webmonitor.JobDetails;
 import org.apache.flink.runtime.rest.handler.legacy.files.StaticFileServerHandler;
-import org.apache.flink.runtime.webmonitor.history.JsonArchivist;
 import org.apache.flink.runtime.webmonitor.retriever.GatewayRetriever;
-import org.apache.flink.runtime.webmonitor.retriever.LeaderGatewayRetriever;
-import org.apache.flink.runtime.webmonitor.retriever.MetricQueryServiceRetriever;
 import org.apache.flink.util.FlinkException;
 
 import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.databind.JsonNode;
@@ -51,7 +44,6 @@ import java.io.File;
 import java.io.IOException;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
 import java.net.URI;
 import java.util.HashMap;
 import java.util.Iterator;
@@ -67,7 +59,7 @@ import java.util.concurrent.Executor;
  */
 public final class WebMonitorUtils {
 
-	private static final String WEB_RUNTIME_MONITOR_CLASS_FQN = "org.apache.flink.runtime.webmonitor.WebRuntimeMonitor";
+	private static final String WEB_FRONTEND_BOOTSTRAP_CLASS_FQN = "org.apache.flink.runtime.webmonitor.utils.WebFrontendBootstrap";
 
 	private static final Logger LOG = LoggerFactory.getLogger(WebMonitorUtils.class);
 
@@ -121,59 +113,6 @@ public final class WebMonitorUtils {
 		private static File resolveFileLocation(String logFilePath) {
 			File logFile = new File(logFilePath);
 			return (logFile.exists() && logFile.canRead()) ? logFile : null;
-		}
-	}
-
-	/**
-	 * Starts the web runtime monitor. Because the actual implementation of the runtime monitor is
-	 * in another project, we load the runtime monitor dynamically.
-	 *
-	 * <p>Because failure to start the web runtime monitor is not considered fatal, this method does
-	 * not throw any exceptions, but only logs them.
-	 *
-	 * @param config The configuration for the runtime monitor.
-	 * @param highAvailabilityServices HighAvailabilityServices used to start the WebRuntimeMonitor
-	 * @param jobManagerRetriever which retrieves the currently leading JobManager
-	 * @param queryServiceRetriever which retrieves the query service
-	 * @param timeout for asynchronous operations
-	 * @param scheduledExecutor to run asynchronous operations
-	 */
-	public static WebMonitor startWebRuntimeMonitor(
-			Configuration config,
-			HighAvailabilityServices highAvailabilityServices,
-			LeaderGatewayRetriever<JobManagerGateway> jobManagerRetriever,
-			MetricQueryServiceRetriever queryServiceRetriever,
-			Time timeout,
-			ScheduledExecutor scheduledExecutor) {
-		// try to load and instantiate the class
-		try {
-			Class<? extends WebMonitor> clazz = Class.forName(WEB_RUNTIME_MONITOR_CLASS_FQN).asSubclass(WebMonitor.class);
-
-			Constructor<? extends WebMonitor> constructor = clazz.getConstructor(
-				Configuration.class,
-				LeaderRetrievalService.class,
-				LeaderGatewayRetriever.class,
-				MetricQueryServiceRetriever.class,
-				Time.class,
-				ScheduledExecutor.class);
-			return constructor.newInstance(
-				config,
-				highAvailabilityServices.getJobManagerLeaderRetriever(HighAvailabilityServices.DEFAULT_JOB_ID),
-				jobManagerRetriever,
-				queryServiceRetriever,
-				timeout,
-				scheduledExecutor);
-		} catch (ClassNotFoundException e) {
-			LOG.error("Could not load web runtime monitor. " +
-					"Probably reason: flink-runtime-web is not in the classpath");
-			LOG.debug("Caught exception", e);
-			return null;
-		} catch (InvocationTargetException e) {
-			LOG.error("WebServer could not be created", e.getTargetException());
-			return null;
-		} catch (Throwable t) {
-			LOG.error("Failed to instantiate web runtime monitor.", t);
-			return null;
 		}
 	}
 
@@ -252,22 +191,6 @@ public final class WebMonitorUtils {
 		} else {
 			throw new FlinkException("The module flink-runtime-web could not be found in the class path. Please add " +
 				"this jar in order to enable web based job submission.");
-		}
-	}
-
-	public static JsonArchivist[] getJsonArchivists() {
-		try {
-			Class<? extends WebMonitor> clazz = Class.forName(WEB_RUNTIME_MONITOR_CLASS_FQN).asSubclass(WebMonitor.class);
-			Method method = clazz.getMethod("getJsonArchivists");
-			return (JsonArchivist[]) method.invoke(null);
-		} catch (ClassNotFoundException e) {
-			LOG.error("Could not load web runtime monitor. " +
-				"Probably reason: flink-runtime-web is not in the classpath");
-			LOG.debug("Caught exception", e);
-			return new JsonArchivist[0];
-		} catch (Throwable t) {
-			LOG.error("Failed to retrieve archivers from web runtime monitor.", t);
-			return new JsonArchivist[0];
 		}
 	}
 
@@ -367,7 +290,7 @@ public final class WebMonitorUtils {
 	 */
 	private static boolean isFlinkRuntimeWebInClassPath() {
 		try {
-			Class.forName(WEB_RUNTIME_MONITOR_CLASS_FQN).asSubclass(WebMonitor.class);
+			Class.forName(WEB_FRONTEND_BOOTSTRAP_CLASS_FQN);
 			return true;
 		} catch (ClassNotFoundException e) {
 			// class not found means that there is no flink-runtime-web in the classpath
